requirements.txt
dash
dash-daq
dash-auth
dash-table
dash-bootstrap-components

Flask-Mail
Flask-Admin
Flask-Login
Flask-Redis
Flask-SQLAlchemy

redis
celery
pymysql
DBUtils
diskcache

uWSGI
orjson
psutil
autopep8
multiprocess

lxml
requests
fake_useragent
beautifulSoup4